tasks are done but w/o mq

+++ src/events/events.service.ts
@@ -1,16 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Between, EntityManager, Repository } from 'typeorm';
 import { EventsDto } from './dto/create-event.input';
 import { EventsEntity } from './entities/event.entity';
-import { LocationEntity } from 'src/locations/entities/location.entity';
+import { LocationEntity } from 'src/events/entities/location.entity';
 import { updEventsDto } from './dto/update-event.dto';
+import { LocationDto } from './dto/create-location.input';
+import { updLocationDto } from './dto/update-location.dto';
 
 @Injectable()
 export class EventsService {
   constructor(
-        @InjectRepository(EventsEntity)
-    private eventsRepo: Repository<EventsEntity>,
+    @InjectRepository(EventsEntity)  private readonly eventsRepo: Repository<EventsEntity>,
+    @InjectRepository(LocationEntity) private readonly locationRepo: Repository<LocationEntity>,
     private readonly entityManager: EntityManager
     ){}
 
@@ -68,48 +70,7 @@ export class EventsService {
         return await this.entityManager.save(newEvent)
     }
 
-    // async updateEvent(id: number, event: updEventsDto): Promise<EventsEntity | string>{
-    //     const updateEvent = await this.eventsRepo.update(id, event)
-    //     if (updateEvent){
-    //     return await this.eventsRepo.findOne({where: {id}, relations: ["location"]})
-    //     }
-    //     else return "not found"
-    // }
-
-    // async updateEvent(id: number, event: updEventsDto){
-    //     await this.entityManager.transaction(async entityManager => {
-    //         const e = await this.eventsRepo.findOne({where: {id}, relations: ["location"]})
-    //         const location = {...event.location}
-    //         // e.location = location
-    //         e.location = {
-    //             id: id,
-    //             ...location
-    //         }
-    //         await entityManager.save(location)
-    //         await entityManager.save(event)
-    //         return this.eventsRepo.findOne({where: {id}, relations: ["location"]})
-    //     })
-    // }
-    // async updateEvent(id: number, event: updEventsDto): Promise<EventsEntity>{
-    //     const location = new LocationEntity({...event.location})
-    //     const e = {
-    //         ...event, 
-    //         location
-    //     }
-    //     await this.eventsRepo.update(id,e)
-    //     return await this.eventsRepo.findOne({where:{id}, relations: ["location"]})
-    // }
-
-    // async updateEvent(id: number, event: updEventsDto): Promise<EventsEntity>{
-    //     let l = new LocationEntity({...event.location})
-    //     const e = {
-    //         ...event, 
-    //         l
-    //     }
-    //     await this.eventsRepo.update(id,event)
-    //     return await this.eventsRepo.findOne({where:{id}, relations: ["location"]})
-    // }
-
+   
     async updateEvent(id: number, event: updEventsDto): Promise<EventsEntity>{
         const f = await this.eventsRepo.findOne({where: {id}, relations: ["location"]})
         return this.eventsRepo.save({id, ...event})
@@ -119,8 +80,76 @@ export class EventsService {
     async deleteEvent(id: number): Promise<EventsEntity | string>   {
         const event = await this.eventsRepo.findOne({where: {id}})
         if (event){
-        await this.eventsRepo.remove(event) 
-        return await event
+            await this.eventsRepo.remove(event) 
+            return await event
+        }
+        return "not found"
+    } 
+
+async getAllLocations(): Promise<EventsEntity[]>{
+        return await this.eventsRepo.find({
+            relations: ["location"]
+        })
+    }
+
+    async getAvailable(): Promise<EventsEntity[] | string>{
+        const available = this.eventsRepo.find({
+            relations: ["location"], 
+            where: {
+              location: {
+                availability: true
+              }
+            }
+        })
+    
+        if (available) {
+            return available
+        }
+        return available
+    }
+
+
+    async getByAddress(address: string): Promise<EventsEntity[] | string> {
+        return await this.eventsRepo.find({
+        relations: ["location"],
+        where: {
+            location: {
+                address : address
+            }
+        }
+        })
+    } 
+
+    async createLocation(location: LocationDto): Promise<LocationEntity>{
+        const newLocation = new LocationEntity(location)
+        return await this.entityManager.save(newLocation)
+    }
+
+    async deleteLocation(id: number) : Promise<LocationEntity | string>{
+        const del = await this.locationRepo.findOne({where: {id}})
+        if (del) {
+            await this.locationRepo.delete(del)
+            return del 
+        }
+        return "not found"
+    }
+
+    async updateLocation(id: number, location: updLocationDto) : Promise<LocationEntity> {
+        const updLocation = await this.locationRepo.update(id, location)
+        return await this.locationRepo.findOne({where: {id}})
+    }
+
+    async deleteByLocation(id: number): Promise<EventsEntity | string>   {
+        const event = await this.eventsRepo.findOne({
+            relations: ["location"],
+            where: {
+                location: {
+                    id: id
+                }
+            }})
+        if (event){
+            await this.eventsRepo.remove(event) 
+            return await event
         }
         return "not found"
     }
